# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Oasis Garden & Patio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/products/import/dry-run:
    post:
      operationId: adminImportProductsDryRun
      tags: [admin-products]
      summary: Validate a CSV import without writing changes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportProductsRequest"
      responses:
        "200":
          description: Dry-run validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportProductsDryRunResult"
        "400":
          description: Validation error (e.g. CSV unparseable, required mapping missing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/products/import/commit:
    post:
      operationId: adminImportProductsCommit
      tags: [admin-products]
      summary: Commit a CSV import (creates or updates products by SKU)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportProductsRequest"
      responses:
        "200":
          description: Commit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportProductsCommitResult"
        "400":
          description: Validation error or rows had blocking errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ImportProductsMapping:
      type: object
      description: Maps product fields to CSV header column names. Required mapped fields - name, sku, manufacturer, category, price.
      properties:
        name: { type: string }
        slug: { type: string }
        sku: { type: string }
        manufacturer: { type: string }
        category: { type: string }
        shortDescription: { type: string }
        description: { type: string }
        price: { type: string }
        cost: { type: string }
        weight: { type: string }
        dimensions: { type: string }
        onHand: { type: string }
        reorderThreshold: { type: string }
        featured: { type: string }
        availableOnline: { type: string }
        inStoreOnly: { type: string }
      required: [name, sku, manufacturer, category, price]
    ImportProductsRequest:
      type: object
      properties:
        csvText:
          type: string
          minLength: 1
        mapping:
          $ref: "#/components/schemas/ImportProductsMapping"
      required: [csvText, mapping]
    ImportProductsRowError:
      type: object
      properties:
        field: { type: string }
        message: { type: string }
      required: [field, message]
    ImportProductsRowResult:
      type: object
      properties:
        rowIndex:
          type: integer
          description: 1-based row number in the CSV (excluding header).
        action:
          type: string
          enum: [create, update, error]
        sku:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        existingProductId:
          type: integer
          nullable: true
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportProductsRowError"
      required: [rowIndex, action, errors]
    ImportProductsDryRunResult:
      type: object
      properties:
        totalRows: { type: integer }
        willCreateCount: { type: integer }
        willUpdateCount: { type: integer }
        errorCount: { type: integer }
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportProductsRowResult"
      required: [totalRows, willCreateCount, willUpdateCount, errorCount, rows]
    ImportProductsCommitResult:
      type: object
      properties:
        totalRows: { type: integer }
        createdCount: { type: integer }
        updatedCount: { type: integer }
        errorCount: { type: integer }
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportProductsRowResult"
      required: [totalRows, createdCount, updatedCount, errorCount, errors]
...